# docs/swagger.yaml
swagger: "2.0"
info:
  title: Restfull API
  description: API hệ thống bán lẻ đồ gia dụng
  version: "1.0"
host: "26.107.218.28:5000"
basePath: "/api"
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Token phải có dạng 'Bearer <JWT>'"
definitions:
...
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
  SuccessResponse:
    type: object
    properties:
      message:
        type: string
paths:
...
  /user/grant-staff/{user_id}:
    put:
      tags:
        - user
      summary: Cấp quyền Staff cho User (Chỉ Admin)
      security:
        - Bearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Cấp quyền Staff thành công
        400:
          description: User đã là Staff hoặc không thể cấp quyền cho Admin
        403:
          description: Không đủ quyền
        404:
          description: User không tồn tại
  /user/update-status/{user_id}:
    put:
      tags:
        - user
      summary: Admin cập nhật trạng thái tài khoản user
      security:
        - Bearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              is_active:
                type: boolean

      responses:
        200:
          description: Trạng thái tài khoản đã được cập nhật
          schema:
            $ref: '#/definitions/SuccessResponse'
        404:
          description: Người dùng không tồn tại
          schema:
            $ref: '#/definitions/ErrorResponse'
        400:
          description: Thiếu trường 'is_active'
          schema:
            $ref: '#/definitions/ErrorResponse'
  /user/profile:
    get:
      tags:
        - user
      summary: Lấy thông tin cá nhân
      security:
        - Bearer: []
      responses:
        200:
          description: Trả về thông tin người dùng
        404:
          description: User không tồn tại
  /user/update-profile:
    put:
      tags:
        - user
      summary: Cập nhật thông tin cá nhân
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              full_name:
                type: string
                required: false
              phone_number:
                type: string
                required: false
              date_of_birth:
                type: string
                format: date
                required: false
              gender:
                type: string
                enum: ["Male", "Female", "Other"]
                required: false
      responses:
        200:
          description: Cập nhật thành công
        400:
          description: Lỗi dữ liệu đầu vào
        404:
          description: User không tồn tại
  /user/update-avatar:
    put:
      tags:
        - user
      summary: Cập nhật ảnh đại diện
      security:
        - Bearer: []
      consumes:
        - multipart/form-data
      parameters:
        - name: avatar
          in: formData
          required: true
          type: file
      responses:
        200:
          description: Cập nhật thành công
        400:
          description: Không có file hoặc lỗi xử lý file
        404:
          description: User không tồn tại
  /user/addresses:
    get:
      tags:
        - user
      summary: Lấy danh sách địa chỉ
      security:
        - Bearer: []
      responses:
        200:
          description: Trả về danh sách địa chỉ của người dùng
        404:
          description: User không tồn tại
        401:
          description: Yêu cầu không hợp lệ hoặc không có quyền truy cập
  /user/add-address:
    post:
      tags:
        - user
      summary: Thêm địa chỉ mới
      security:
        - Bearer: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              full_name:
                type: string
              phone_number:
                type: string
              street_address:
                type: string
              district:
                type: string
              province:
                type: string
              city:
                type: string
              state:
                type: string
              country:
                type: string
              postal_code:
                type: string
              is_default:
                type: boolean
      responses:
        201:
          description: Địa chỉ được thêm thành công
        400:
          description: Dữ liệu không hợp lệ
        401:
          description: Yêu cầu không hợp lệ hoặc không có quyền truy cập
  /user/update-address/{address_id}:
    put:
      tags:
        - user
      summary: Cập nhật địa chỉ
      security:
        - Bearer: []
      parameters:
        - name: address_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              full_name:
                type: string
              phone_number:
                type: string
              street_address:
                type: string
              district:
                type: string
              province:
                type: string
              city:
                type: string
              state:
                type: string
              country:
                type: string
              postal_code:
                type: string
              is_default:
                type: boolean
      responses:
        200:
          description: Đã cập nhật địa chỉ thành công
        404:
          description: Địa chỉ không tồn tại
          schema:
            $ref: '#/definitions/ErrorResponse'
  /user/delete-address/{address_id}:
    delete:
      tags:
        - user
      summary: Xóa địa chỉ
      security:
        - Bearer: []
      parameters:
        - name: address_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Xóa thành công
        404:
          description: Địa chỉ không tồn tại
        401:
          description: Yêu cầu không hợp lệ hoặc không có quyền truy cập
  /user/set-default-address/{address_id}:
    put:
      tags:
        - user
      summary: Đặt địa chỉ mặc định
      security:
        - Bearer: []
      parameters:
        - name: address_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: Đã đặt địa chỉ mặc định
        404:
          description: Địa chỉ không tồn tại
        401:
          description: Yêu cầu không hợp lệ hoặc không có quyền truy cập
  /user/users:
    get:
      tags:
        - user
      summary: Lấy danh sách user (Chỉ Admin)
      security:
        - Bearer: []
      parameters:
        - name: role
          in: query
          required: false
          type: string
          description: "Lọc theo vai trò người dùng (Admin, Staff, Customer)"
        - name: page
          in: query
          required: false
          type: integer
          description: "Số trang hiện tại"
        - name: limit
          in: query
          required: false
          type: integer
          description: "Số lượng user trên mỗi trang"
      responses:
        200:
          description: Danh sách người dùng
          schema:
            type: object
            properties:
              total:
                type: integer
              page:
                type: integer
              per_page:
                type: integer
              users:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    username:
                      type: string
                    email:
                      type: string
                    full_name:
                      type: string
                    role:
                      type: string
                    avatar_url:
                      type: string
                    is_active:
                      type: boolean
        403:
          description: Không đủ quyền
        401:
          description: Yêu cầu không hợp lệ hoặc không có quyền truy cập
  /user/delete/{user_id}:
    delete:
      tags:
        - user
      summary: Xóa người dùng (Chỉ Admin)
      description: Cho phép admin xóa người dùng theo ID. Không thể xóa tài khoản admin.
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID của người dùng cần xóa
          schema:
            type: integer
      responses:
        '200':
          description: Xóa người dùng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User deleted successfully
        '403':
          description: Không thể xóa tài khoản admin
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Cannot delete admin account
        '404':
          description: Không tìm thấy người dùng
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User does not exist
...